Refactor: move commonly used functions to common package
