+++ .gitignore
@@ -3,11 +3,11 @@ image_builder/deploy/*
 k8s/migration/dist/*
 k8s/migration/vendor/*
 v2v-helper/vendor/*
-v2v-helper/manager
 pkg/vpwned/vendor/*
+pkg/common/vendor/*
 k8s/migration/manager
+v2v-helper/manager
 bin/*
-pkg/vpwned/vendor/*
 bin/*
 .vite
 .env
@@ -16,6 +16,5 @@ bin/*
 node_modules
 openstack.rc
 .qodo
-pkg/vpwned/vendor/*
 pkg/vpwned/bin/vpwctl
 proto_dependencies
