+++ pkg/common/k8s/client.go
@@ -1,0 +1,33 @@
+package k8s
+
+import (
+	"github.com/pkg/errors"
+	vjailbreakv1alpha1 "github.com/platform9/vjailbreak/k8s/migration/api/v1alpha1"
+	"k8s.io/apimachinery/pkg/runtime"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+// GetInclusterClient creates a Kubernetes client configured for in-cluster use
+// with the vjailbreak scheme registered.
+// This function is used by both k8s/migration and v2v-helper modules.
+func GetInclusterClient() (client.Client, error) {
+	// Create a direct Kubernetes client
+	config, err := rest.InClusterConfig()
+	if err != nil {
+		return nil, errors.Wrap(err, "Failed to get in-cluster config")
+	}
+	scheme := runtime.NewScheme()
+	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
+	utilruntime.Must(vjailbreakv1alpha1.AddToScheme(scheme))
+	clientset, err := client.New(config, client.Options{
+		Scheme: scheme,
+	})
+	if err != nil {
+		return nil, errors.Wrap(err, "Failed to get in-cluster config")
+	}
+
+	return clientset, err
+}
