+++ v2v-helper/pkg/utils/utils.go
@@ -9,14 +9,10 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
-	vjailbreakv1alpha1 "github.com/platform9/vjailbreak/k8s/migration/api/v1alpha1"
 	"github.com/platform9/vjailbreak/pkg/common/constants"
+	k8scommon "github.com/platform9/vjailbreak/pkg/common/k8s"
 	"github.com/platform9/vjailbreak/v2v-helper/pkg/k8sutils"
 
-	"k8s.io/apimachinery/pkg/runtime"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
 	client "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -31,22 +27,7 @@ func RemoveEmptyStrings(slice []string) []string {
 }
 
 func GetInclusterClient() (client.Client, error) {
-	// Create a direct Kubernetes client
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to get in-cluster config")
-	}
-	scheme := runtime.NewScheme()
-	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-	utilruntime.Must(vjailbreakv1alpha1.AddToScheme(scheme))
-	clientset, err := client.New(config, client.Options{
-		Scheme: scheme,
-	})
-	if err != nil {
-		return nil, errors.Wrap(err, "Failed to get in-cluster config")
-	}
-
-	return clientset, err
+	return k8scommon.GetInclusterClient()
 }
 
 func PrintLog(logMessage string) error {
